// src/api/models/tasks-stats-list200-response-categories-inner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Your API
 * API description
 *
 * The version of the OpenAPI document: v1
 * Contact: dev@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TasksStatsList200ResponseCategoriesInner
 */
export interface TasksStatsList200ResponseCategoriesInner {
    /**
     * 
     * @type {string}
     * @memberof TasksStatsList200ResponseCategoriesInner
     */
    'name'?: string;
    /**
     * 
     * @type {number}
     * @memberof TasksStatsList200ResponseCategoriesInner
     */
    'total'?: number;
    /**
     * 
     * @type {number}
     * @memberof TasksStatsList200ResponseCategoriesInner
     */
    'completed'?: number;
    /**
     * 
     * @type {number}
     * @memberof TasksStatsList200ResponseCategoriesInner
     */
    'value'?: number;
    /**
     * 
     * @type {string}
     * @memberof TasksStatsList200ResponseCategoriesInner
     */
    'color'?: string;
}